Refactor Quest Book structure
