+++ common/src/main/java/com/wynntils/wynn/event/QuestBookReloadedEvent.java
@@ -1,0 +1,9 @@
+/*
+ * Copyright © Wynntils 2022.
+ * This file is released under AGPLv3. See LICENSE for full license details.
+ */
+package com.wynntils.wynn.event;
+
+import net.minecraftforge.eventbus.api.Event;
+
+public class QuestBookReloadedEvent extends Event {}
